[Refactor] Extension 파일 네이밍 수정 & 그룹화
